// src/_shared/app/enum/rule.enum.ts
export enum RuleEnum {
  // APP TASK SETTING
  CAN_SHOW_SETTING = 'CAN_SHOW_SETTING',
  // DOSSIER
  CAN_CREATE_DOSSIER = 'CAN_CREATE_DOSSIER',
  CAN_SHOW_DOSSIER = 'CAN_SHOW_DOSSIER',
  CAN_SHOW_DOSSIER_LIST = 'CAN_SHOW_DOSSIER_LIST',
  CAN_UPDATE_DOSSIER = 'CAN_UPDATE_DOSSIER',
  CAN_DELETE_DOSSIER = 'CAN_DELETE_DOSSIER',
  // AFFECTATION
  CAN_SHOW_AFFECTATION = 'CAN_SHOW_AFFECTATION',
  CAN_SHOW_AFFECTATION_LIST = 'CAN_SHOW_AFFECTATION_LIST',
  CAN_UPDATE_AFFECTATION = 'CAN_UPDATE_AFFECTATION',
  
  // PATRIMOINE
  CAN_SHOW_PATRIMOINE = 'CAN_SHOW_PATRIMOINE',
  CAN_SHOW_PATRIMOINE_LIST = 'CAN_SHOW_PATRIMOINE_LIST',
  // DOSSIER_STEP
  CAN_SHOW_DOSSIER_STEP_LIST = 'CAN_SHOW_DOSSIER_STEP_LIST',
  CAN_SHOW_OWN_DOSSIER_STEP_LIST = 'CAN_SHOW_OWN_DOSSIER_STEP_LIST',
  // CLIENT
  CAN_SHOW_CLIENT_LIST = 'CAN_SHOW_CLIENT_LIST',
  CAN_SHOW_CLIENT = 'CAN_SHOW_CLIENT',
  // TASK
  CAN_SHOW_TASK = 'CAN_SHOW_TASK',
  CAN_SHOW_TASK_LIST = 'CAN_SHOW_TASK_LIST'
}